refactor(settings): split settings views into focused files
