+++ notchi/notchi/Views/SettingsRowViews.swift
@@ -1,0 +1,62 @@
+import SwiftUI
+
+struct SettingsStatusBadge: View {
+    let text: String
+    let color: Color
+
+    var body: some View {
+        Text(text)
+            .font(.system(size: 10, weight: .medium))
+            .foregroundColor(color)
+            .lineLimit(1)
+            .truncationMode(.tail)
+            .padding(.horizontal, 6)
+            .padding(.vertical, 2)
+            .background(color.opacity(0.15))
+            .cornerRadius(4)
+            .frame(maxWidth: 160, alignment: .trailing)
+    }
+}
+
+struct SettingsRowView<Trailing: View>: View {
+    let icon: String
+    let title: LocalizedStringKey
+    @ViewBuilder let trailing: () -> Trailing
+
+    var body: some View {
+        HStack {
+            Image(systemName: icon)
+                .font(.system(size: 12))
+                .foregroundColor(TerminalColors.secondaryText)
+                .frame(width: 20)
+
+            Text(title)
+                .font(.system(size: 12))
+                .foregroundColor(TerminalColors.primaryText)
+
+            Spacer()
+
+            trailing()
+        }
+        .padding(.vertical, SettingsLayout.rowVerticalPadding)
+        .contentShape(Rectangle())
+    }
+}
+
+struct ToggleSwitch: View {
+    let isOn: Bool
+
+    var body: some View {
+        ZStack(alignment: isOn ? .trailing : .leading) {
+            Capsule()
+                .fill(isOn ? TerminalColors.green : Color.white.opacity(0.15))
+                .frame(width: 32, height: 18)
+
+            Circle()
+                .fill(Color.white)
+                .frame(width: 14, height: 14)
+                .padding(2)
+        }
+        .animation(.easeInOut(duration: 0.15), value: isOn)
+    }
+}
